3D Engine  + Fully implemented and tested external texture system Map Builder 2.0 and Texture Compiler  + Fixed minor issues (ie. speed should have been a double)
git-svn-id: file:///e/SVN/3d_engine@41 463a7de7-ff88-454d-8a36-9b7cb3bdd095

+++ Builder 2.0/src/AnimatedTexture.java
@@ -13,9 +13,9 @@ public class AnimatedTexture {
 	public String name;
 	public int firstTextureIndex;
 	public int frames;
-	public int speed;
+	public double speed;
 	
-	public AnimatedTexture(int index, String name, String firstTextureName, int frames, int speed, Vector<String> textureNames) {
+	public AnimatedTexture(int index, String name, String firstTextureName, int frames, double speed, Vector<String> textureNames) {
 		this.index = index;
 		this.name = name;
 		this.frames = frames;
@@ -73,7 +73,7 @@ public class AnimatedTexture {
 				this.frames = Integer.valueOf(newProperty.value);
 			}
 			else if(newProperty.key.equalsIgnoreCase("speed")) {
-				this.speed = Integer.valueOf(newProperty.value);
+				this.speed = Double.valueOf(newProperty.value);
 			}
 		}
 	}
